Quick and easy container node ssh
